spike: rename -> splocator
